# frontend/docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Orient-advertise-backend API",
    "description" : "Multi-module Spring Boot platform for TV-box content delivery — device management, content uploads + transcoding, playlists, scheduling, remote actions, incident lifecycle, proof-of-play, aggregated reporting, Excel export, and external partner integrations.\n\n**Authentication.** Most endpoints require a JWT access token (`Authorization: Bearer <token>`) obtained from `POST /auth/login`. External integration endpoints under `/service/external/**` use an `X-API-Key` header instead — see the *External* group.\n\n**Sensitive endpoints.** Operations marked with `[SENSITIVE]` in their summary affect production state, expose user data, or irreversibly delete records. Audit logging is enabled and the operator's username is recorded.\n",
    "contact" : {
      "name" : "orientadvertise.services"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version" : "1.0.55"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Current host"
  }, {
    "url" : "https://api.example.com",
    "description" : "Production"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/devices/{id}/playback" : {
      "post" : {
        "tags" : [ "playback-controller" ],
        "operationId" : "record",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/JsonNode"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BatchResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/reset-password" : {
      "get" : {
        "tags" : [ "Auth" ],
        "summary" : "Check whether a reset token is still valid",
        "description" : "Non-consuming. Public token oracle, so it is rate-limited the same as the reset POST to blunt brute force (the 256-bit token makes guessing infeasible regardless).",
        "operationId" : "validateToken",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "429" : {
            "description" : "Rate limited",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidateTokenResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "{ valid: boolean }",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidateTokenResponse"
                }
              }
            }
          }
        },
        "security" : [ ]
      },
      "post" : {
        "tags" : [ "Auth" ],
        "summary" : "Reset a password using a one-time token",
        "operationId" : "resetPassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "429" : {
            "description" : "Rate limited"
          },
          "400" : {
            "description" : "Bad/expired/used token, mismatch, or policy"
          },
          "204" : {
            "description" : "Password reset; all sessions revoked"
          }
        },
        "security" : [ ]
      }
    },
...
    "/api/auth/forgot-password" : {
      "post" : {
        "tags" : [ "Auth" ],
        "summary" : "Request a password-reset link",
        "description" : "If an active account owns the email, a one-time reset link is emailed. The response is identical (202) whether or not the email exists — no account enumeration. Rate-limited per IP and per email.",
        "operationId" : "forgotPassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "429" : {
            "description" : "Rate limited"
          },
          "400" : {
            "description" : "Malformed email"
          },
          "202" : {
            "description" : "Accepted (always, regardless of whether the email exists)"
          }
        },
        "security" : [ ]
      }
    },
...
    "/api/sync-groups/{id}/playback" : {
      "get" : {
        "tags" : [ "Sync Groups" ],
        "summary" : "Sync-group playback view (pickable items + active jump)",
        "description" : "The shared deliverable timeline the whole group can jump to. When the members are not content-coherent (different playlists/versions, a member with no active playlist, or an empty group) `coherent` is false with a `reason` and an empty item list.",
        "operationId" : "playback",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Unknown id, or out of operator scope",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SyncGroupPlaybackView"
                }
              }
            }
          },
          "403" : {
            "description" : "Caller lacks ADMIN/OPERATOR/VIEWER",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SyncGroupPlaybackView"
                }
              }
            }
          },
          "200" : {
            "description" : "Playback view (coherent, or coherent:false + reason)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SyncGroupPlaybackView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/devices" : {
      "get" : {
        "tags" : [ "device-controller" ],
        "operationId" : "list_10",
        "parameters" : [ {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "ONLINE", "OFFLINE", "NO_CONTENT", "UNREGISTERED" ]
          }
        }, {
          "name" : "regionId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "projectId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "facilityId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "deviceGroupId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "unassigned",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "serial",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "name",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "facilityName",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "hasActivePlaylist",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "syncUnassigned",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageDeviceListItem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "JsonNode" : {
        "type" : "object"
      },
      "BatchResponse" : {
        "type" : "object",
        "properties" : {
          "total" : {
            "type" : "integer",
            "format" : "int32"
          },
          "created" : {
            "type" : "integer",
            "format" : "int32"
          },
          "duplicate" : {
            "type" : "integer",
            "format" : "int32"
          },
          "rejected" : {
            "type" : "integer",
            "format" : "int32"
          },
          "rejections" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RejectionDto"
            }
          }
        }
      },
      "RejectionDto" : {
        "type" : "object",
        "properties" : {
          "index" : {
            "type" : "integer",
            "format" : "int32"
          },
          "reason" : {
            "type" : "string"
          }
        }
      },
...
      "ResetPasswordRequest" : {
        "required" : [ "confirmPassword", "newPassword", "token" ],
        "type" : "object",
        "properties" : {
          "token" : {
            "type" : "string"
          },
          "newPassword" : {
            "maxLength" : 200,
            "minLength" : 8,
            "type" : "string"
          },
          "confirmPassword" : {
            "type" : "string"
          }
        }
      },
...
      "ForgotPasswordRequest" : {
        "required" : [ "email" ],
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          }
        }
      },
...
      "Pageable" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "PageableObject" : {
        "type" : "object",
        "properties" : {
          "offset" : {
            "type" : "integer",
            "format" : "int64"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "unpaged" : {
            "type" : "boolean"
          },
          "paged" : {
            "type" : "boolean"
          }
        }
      },
      "SortObject" : {
        "type" : "object",
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        }
      },
...
      "ActiveJump" : {
        "type" : "object",
        "properties" : {
          "index" : {
            "type" : "integer",
            "format" : "int32"
          },
          "activateAt" : {
            "type" : "integer",
            "format" : "int64"
          },
          "activateAtIso" : {
            "type" : "string"
          }
        }
      },
      "SyncGroupPlaybackItem" : {
        "type" : "object",
        "properties" : {
          "index" : {
            "type" : "integer",
            "format" : "int32"
          },
          "fileId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "title" : {
            "type" : "string"
          },
          "durationSeconds" : {
            "type" : "integer",
            "format" : "int32"
          },
          "slotStartMs" : {
            "type" : "integer",
            "format" : "int64"
          },
          "slotDurationMs" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "SyncGroupPlaybackView" : {
        "type" : "object",
        "properties" : {
          "syncGroupId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "coherent" : {
            "type" : "boolean"
          },
          "reason" : {
            "type" : "string"
          },
          "playlistId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "playlistName" : {
            "type" : "string"
          },
          "loopDurationMs" : {
            "type" : "integer",
            "format" : "int64"
          },
          "memberCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SyncGroupPlaybackItem"
            }
          },
          "activeJump" : {
            "$ref" : "#/components/schemas/ActiveJump"
          }
        }
      },
...
      "DeviceListItem" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "serialNumber" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "computedStatus" : {
            "type" : "string"
          },
          "regionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "facilityId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "facilityName" : {
            "type" : "string"
          },
          "deviceGroupId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "syncGroupId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lastHeartbeatAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "activePlaylistId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "activePlaylistName" : {
            "type" : "string"
          }
        }
      },
      "PageDeviceListItem" : {
        "type" : "object",
        "properties" : {
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DeviceListItem"
            }
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
...
      "ValidateTokenResponse" : {
        "type" : "object",
        "properties" : {
          "valid" : {
            "type" : "boolean"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "JWT access token from `POST /auth/login`. Roles encoded in the token: ADMIN, OPERATOR, VIEWER, ADVERTISER.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "apiKeyAuth" : {
        "type" : "apiKey",
        "description" : "External-integration API key. Issued by ADMIN via `POST /service/admin/service-keys`. Rate-limited 100 req/hr per key. Revocation is immediate — no in-memory cache widens the window.",
        "name" : "X-API-Key",
        "in" : "header"
      }
    }
  }
}